+++ semcore/carousel/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [3.42.2] - 2024-08-05
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.32.1 ~> 4.32.2], `@semcore/button` [5.34.1 ~> 5.35.0], `@semcore/modal` [4.42.1 ~> 4.43.0], `@semcore/breakpoints` [1.31.1 ~> 1.31.2]).
+
+## [3.42.1] - 2024-07-30
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.32.0 ~> 4.32.1], `@semcore/button` [5.34.0 ~> 5.34.1], `@semcore/breakpoints` [1.31.0 ~> 1.31.1]).
+
+## [3.42.0] - 2024-07-26
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/utils` [4.31.0 ~> 4.32.0], `@semcore/button` [5.33.0 ~> 5.34.0], `@semcore/modal` [4.41.0 ~> 4.42.0], `@semcore/breakpoints` [1.30.0 ~> 1.31.0]).
+
+## [3.41.0] - 2024-07-17
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/button` [5.32.0 ~> 5.33.0]).
+
 ## [3.40.0] - 2024-07-13
 
 ### Changed