+++ semcore/date-picker/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.50.2] - 2024-08-05
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.32.1 ~> 4.32.2], `@semcore/base-trigger` [4.42.1 ~> 4.42.2], `@semcore/button` [5.34.1 ~> 5.35.0], `@semcore/dropdown` [4.35.1 ~> 4.36.0], `@semcore/checkbox` [7.34.1 ~> 7.34.2], `@semcore/tooltip` [6.40.1 ~> 6.40.2], `@semcore/core` [2.29.1 ~> 2.29.2]).
+
+## [4.50.1] - 2024-07-30
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.32.0 ~> 4.32.1], `@semcore/base-trigger` [4.42.0 ~> 4.42.1], `@semcore/button` [5.34.0 ~> 5.34.1], `@semcore/checkbox` [7.34.0 ~> 7.34.1], `@semcore/core` [2.29.0 ~> 2.29.1]).
+
+## [4.50.0] - 2024-07-26
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/utils` [4.31.0 ~> 4.32.0], `@semcore/base-trigger` [4.41.0 ~> 4.42.0], `@semcore/button` [5.33.0 ~> 5.34.0], `@semcore/divider` [4.29.0 ~> 4.30.0], `@semcore/checkbox` [7.33.0 ~> 7.34.0], `@semcore/tooltip` [6.39.1 ~> 6.40.0], `@semcore/core` [2.28.0 ~> 2.29.0]).
+
+## [4.49.0] - 2024-07-17
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/button` [5.32.0 ~> 5.33.0], `@semcore/tooltip` [6.37.0 ~> 6.39.0]).
+
 ## [4.48.0] - 2024-07-13
 
 ### Changed