+++ semcore/fullscreen-modal/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [3.43.0] - 2024-07-29
+
+### Changed
+
+- Close icon to `Button` component.
+
+## [3.42.2] - 2024-08-05
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.32.1 ~> 4.32.2], `@semcore/modal` [4.42.1 ~> 4.43.0], `@semcore/flex-box` [5.31.1 ~> 5.31.2], `@semcore/core` [2.29.1 ~> 2.29.2]).
+
+## [3.42.1] - 2024-07-30
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.32.0 ~> 4.32.1], `@semcore/flex-box` [5.31.0 ~> 5.31.1], `@semcore/core` [2.29.0 ~> 2.29.1]).
+
+## [3.42.0] - 2024-07-26
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/utils` [4.31.0 ~> 4.32.0], `@semcore/modal` [4.41.0 ~> 4.42.0], `@semcore/flex-box` [5.30.0 ~> 5.31.0], `@semcore/core` [2.28.0 ~> 2.29.0]).
+
 ## [3.41.0] - 2024-07-13
 
 ### Changed