+++ semcore/scroll-area/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.35.0] - 2024-07-31
+
+### Add
+
+- Context to hide scrollbars from screen readers.
+
+## [5.34.1] - 2024-07-30
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.32.0 ~> 4.32.1], `@semcore/flex-box` [5.31.0 ~> 5.31.1], `@semcore/core` [2.29.0 ~> 2.29.1]).
+
+## [5.34.0] - 2024-07-22
+
+### Added
+
+- Property `observeParentSize` to enable recalculation if the parent of ScrollArea is resized. Default value is `false`.
+
+## [5.33.1] - 2024-07-22
+
+### Fixed
+
+- Glitches if the parent of scrollArea has a decimal height or width value.
+
 ## [5.33.0] - 2024-07-13
 
 ### Changed