+++ semcore/time-picker/CHANGELOG.md
@@ -2,6 +2,38 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.47.0] - 2024-07-29
+
+### Changed
+
+- Fixed reading of dropdowns by Screen readers.
+- Removed unnecessary `aria-hidden` props.
+- Text in some `aria-label`.
+
+## [4.46.1] - 2024-07-30
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/button` [5.34.0 ~> 5.34.1], `@semcore/flex-box` [5.31.0 ~> 5.31.1], `@semcore/input` [4.32.0 ~> 4.32.1], `@semcore/select` [4.45.0 ~> 4.45.1], `@semcore/utils` [4.32.0 ~> 4.32.1], `@semcore/core` [2.29.0 ~> 2.29.1]).
+
+## [4.46.0] - 2024-07-26
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/button` [5.33.0 ~> 5.34.0], `@semcore/flex-box` [5.30.0 ~> 5.31.0], `@semcore/input` [4.31.0 ~> 4.32.0], `@semcore/select` [4.44.5 ~> 4.45.0], `@semcore/utils` [4.31.0 ~> 4.32.0], `@semcore/core` [2.28.0 ~> 2.29.0]).
+
+## [4.45.1] - 2024-07-19
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/select` [4.44.4 ~> 4.44.5]).
+
+## [4.45.0] - 2024-07-17
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/button` [5.32.0 ~> 5.33.0]).
+
 ## [4.44.4] - 2024-07-13
 
 ### Changed